refactor: AbortSession and IsSessionAbortedOrUnavailable usecases
